Revert "feat(profile): align banner with space chrome and sampling accents"
This reverts commit cc1ee4e3b21e27689334163d1f41c1f9237028f9.
