+++ apps/web/src/app/[lang]/profile/[personSlug]/page.tsx
@@ -1,6 +1,7 @@
 import {
   EscrowDepositBanners,
   MFABanner,
+  PersonHead,
   ProfileMemberSpaces,
   ProfilePageParams,
 } from '@hypha-platform/epics';
@@ -8,24 +9,18 @@ import Link from 'next/link';
 import { ChevronLeftIcon } from '@radix-ui/react-icons';
 import { Text } from '@radix-ui/themes';
 import { Container, Separator } from '@hypha-platform/ui';
+import React from 'react';
 import {
   findPersonBySlug,
   getSpacesByWeb3Ids,
 } from '@hypha-platform/core/server';
 import { db } from '@hypha-platform/storage-postgres';
-import {
-  getMemberSpaces,
-  Space,
-  DEFAULT_SPACE_AVATAR_IMAGE,
-  DEFAULT_SPACE_LEAD_IMAGE,
-} from '@hypha-platform/core/client';
+import { getMemberSpaces, Space } from '@hypha-platform/core/client';
 import { ProfileTabs } from './_components/profile-tabs';
 import { web3Client } from '@hypha-platform/core/server';
 import { Hex, zeroAddress } from 'viem';
 import { tryDecodeUriPart } from '@hypha-platform/ui-utils';
 import { getTranslations } from 'next-intl/server';
-import { ProfileChromeClient } from './_components/profile-chrome-client';
-import './profile-accent.css';
 
 type PageProps = {
   params: Promise<ProfilePageParams>;
@@ -60,29 +55,6 @@ export default async function ProfilePage(props: PageProps) {
     console.error('Failed to fetch member spaces:', error);
   }
 
-  const rawLead = person?.leadImageUrl?.trim();
-  const heroBannerImageHref =
-    person &&
-    rawLead &&
-    ((rawLead.startsWith('/') && !rawLead.startsWith('//')) ||
-      /^https?:\/\//i.test(rawLead))
-      ? rawLead
-      : DEFAULT_SPACE_LEAD_IMAGE;
-
-  const rawAvatar = person?.avatarUrl?.trim();
-  const accentLogoHref =
-    person &&
-    rawAvatar &&
-    ((rawAvatar.startsWith('/') && !rawAvatar.startsWith('//')) ||
-      /^https?:\/\//i.test(rawAvatar))
-      ? rawAvatar
-      : DEFAULT_SPACE_AVATAR_IMAGE;
-
-  const displayName = person
-    ? `${person.name ?? ''} ${person.surname ?? ''}`.trim()
-    : '';
-  const logoAlt = displayName || person?.slug || 'Profile';
-
   return (
     <Container className="w-full">
       <div className="mb-6 flex items-center w-full">
@@ -98,46 +70,33 @@ export default async function ProfilePage(props: PageProps) {
         </Text>
       </div>
       {person ? (
-        <>
-          {heroBannerImageHref ? (
-            <link
-              rel="preload"
-              as="image"
-              href={heroBannerImageHref}
-              fetchPriority="high"
-            />
-          ) : null}
-          <ProfileChromeClient
-            heroBannerImageHref={heroBannerImageHref}
-            accentLogoHref={accentLogoHref}
-            displayName={displayName || person.slug || ''}
-            logoAlt={logoAlt}
-            description={person.description ?? ''}
-            links={(person.links ?? []).filter(
-              (l): l is string => typeof l === 'string',
-            )}
-            email={person.email ?? ''}
-            location={person.location ?? ''}
-            slug={person.slug ?? ''}
-            createdAt={person.createdAt ?? null}
+        <div className="flex flex-col gap-5">
+          <PersonHead
+            avatar={person?.avatarUrl ?? '/placeholder/default-profile.svg'}
+            name={person?.name ?? ''}
+            surname={person?.surname ?? ''}
+            background={person?.leadImageUrl ?? ''}
+            links={person?.links ?? []}
+            about={person?.description ?? ''}
+            location={person?.location ?? ''}
+            email={person?.email ?? ''}
+            slug={person?.slug ?? ''}
+            createdAt={person?.createdAt}
             exportEmbeddedWallet={true}
-          >
-            <div className="flex flex-col gap-5">
-              <Separator />
-              <MFABanner />
-              <EscrowDepositBanners
-                personSlug={person.slug ?? ''}
-                personAddress={(person.address as `0x${string}`) ?? null}
-              />
-              <ProfileMemberSpaces
-                person={person}
-                spaces={spaces}
-                profileView={true}
-              />
-              <ProfileTabs person={person} lang={lang} />
-            </div>
-          </ProfileChromeClient>
-        </>
+          />
+          <Separator />
+          <MFABanner />
+          <EscrowDepositBanners
+            personSlug={person.slug ?? ''}
+            personAddress={(person.address as `0x${string}`) ?? null}
+          />
+          <ProfileMemberSpaces
+            person={person}
+            spaces={spaces}
+            profileView={true}
+          />
+          <ProfileTabs person={person} lang={lang} />
+        </div>
       ) : (
         <p>{tProfile('personNotFound')}</p>
       )}
