--- apps/web/src/app/[lang]/profile/[personSlug]/profile-accent.css
@@ -1,11 +1,0 @@
-/**
- * Profile accent scope — same CSS variable contract as space pages (`SpaceAccentFromImages`).
- */
-[data-space-accent-scope] {
-  --space-tab-active-border: var(--space-accent, #3b82f6);
-}
-
-[data-space-accent-scope] .tabs-root [data-state='active'] {
-  border-bottom-color: var(--space-tab-active-border) !important;
-  color: var(--color-accent-11, #4a65d8) !important;
-}
