+++ packages/epics/src/common/index.ts
@@ -29,7 +29,3 @@ export {
   HumanChatPanelMessages,
 } from './human-chat-panel';
 export { HumanRightPanel } from './human-right-panel';
-export {
-  StickyAccentChrome,
-  type StickyAccentChromeProps,
-} from './sticky-accent-chrome';
